+++ WebUI/classes/EntryElement.js
@@ -28,8 +28,8 @@ const EntryElement = function EntryElement() {
 		
 		switch(_type) {
 			case EntryType.Boolean:
-				let yes	= BotEditor.I18N('Yes');
-				let no	= BotEditor.I18N('No');
+				let yes	= 'Yes';
+				let no	= 'No';
 				
 				_container.appendChild(arrow_left);
 				_container.appendChild(this._createText(_value == null ? (_default == null ? '' : (_default ? yes : no)) : (_value ? yes : no)));
@@ -156,8 +156,8 @@ const EntryElement = function EntryElement() {
 
 		switch(_type) {
 			case EntryType.Boolean:
-				let yes	= BotEditor.I18N('Yes');
-				let no	= BotEditor.I18N('No');
+				let yes	= 'Yes';
+				let no	= 'No';
 				
 				_container.querySelector('ui-text').innerHTML = (_value ? yes : no);
 			break;