+++ ext/Server/UIServer.lua
@@ -377,10 +377,10 @@ function FunBotUIServer:_writeSettings(p_Player, p_Request)
 		end
 	end
 
-	--UI
+	-- Language of UI
 	if updateLanguage then
-		NetEvents:SendTo('UI_Change_Language', p_Player, p_Request.language)
-		Language:loadLanguage(p_Request.language)
+		Language:loadLanguage(Config.Language)
+		NetEvents:SendTo('UI_Change_Language', p_Player, Config.Language)
 	end
 
 	-- Call batched process