+++ .github/CHANGELOG.md
@@ -27,7 +27,10 @@ This release contains a lot of small bug fixes and more QoL updates.
 * Improved Code of Bot.lua to separate different Functions and for future functions
 * Improved Nade calculations (still far from perfect)
 * land-vehicles teleport if stuck
-* Added Language-System again
+* Added Language-System again! You can now create your own translation
+	* fill out the file ext\Shared\Languages\DEFAULT.lua
+	* fill out the file WebUI\languages\DEFAULT.js 
+	* send those files to me and we will add your language to fun-bots
 
 ### 🐛 Minor bug fixes
 * fixed bug with jeeps
@@ -38,6 +41,9 @@ This release contains a lot of small bug fixes and more QoL updates.
 * Added FOV-Settings for Vehicles
 * fixed Garbage-Collection on Server
 * fixed some vehicle-categories
+* fixed bug in path-switching
+* fixed a bug in target-objective-finding
+* fixed some wrong objective-names on existing maps
 
 ### New maps
 * Theran Highway CQS (Vehicles) - thanks to Gemini899
@@ -59,6 +65,8 @@ This release contains a lot of small bug fixes and more QoL updates.
 * Wake Island Assaullt Small (1+2) (infantery) - thanks to DuTcHrEaGaN
 * Wake Island Assault Large (infantery) - thanks to DuTcHrEaGaN
 * Wake Island Assault Large (Vehicles) - thanks to KrazyIvan777
+* Death Valley CQS (Vehicles) - thanks to KrazyIvan777
+* Riverside CQS (Vehicles) - thanks to KrazyIvan777
 
 ### 📖 Documentation update
 * nothing done on this front :-(